# admin/openapi-films.json
{"openapi":"3.1.0","info":{"title":"Movie theater","description":"Информация о фильмах, жанрах и людях, участвовавших в создании произведения","version":"1.0.0"},"paths":{"/api/v1/films/{film_id}":{"get":{"tags":["Кинопроизведения"],"summary":"Получить информацию о кинопроизведении","description":"Возвращает полную информацию о кинопроизведении по его идентификатору","operationId":"film_details_api_v1_films__film_id__get","parameters":[{"name":"film_id","in":"path","required":true,"schema":{"type":"string","description":"UUID фильма","title":"Film Id"},"description":"UUID фильма","example":"3d825f60-9fff-4dfe-b294-1a45fa1e115d"}],"responses":{"200":{"description":"Объект с информацией о кинопроизведении","content":{"application/json":{"schema":{"$ref":"#/components/schemas/FilmFullResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["Кинопроизведения"],"summary":"Delete Film","operationId":"delete_film_api_v1_films__film_id__delete","security":[{"HTTPBearer":[]}],"parameters":[{"name":"film_id","in":"path","required":true,"schema":{"type":"string","title":"Film Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/films/":{"get":{"tags":["Кинопроизведения"],"summary":"Получить список фильмов по заданным критериям","description":"Возвращает список из короткой информации по кинопроизведениям по заданным фильтрам","operationId":"list_films_api_v1_films__get","parameters":[{"name":"sort","in":"query","required":false,"schema":{"type":"string","pattern":"^-?imdb_rating$","description":"Сортировка (-imdb_rating для DESC, imdb_rating для ASC)","default":"-imdb_rating","title":"Sort"},"description":"Сортировка (-imdb_rating для DESC, imdb_rating для ASC)"},{"name":"genre","in":"query","required":false,"schema":{"anyOf":[{"type":"string","format":"uuid"},{"type":"null"}],"description":"UUID жанра для фильтрации","title":"Genre"},"description":"UUID жанра для фильтрации"},{"name":"page","in":"query","required":false,"schema":{"type":"integer","minimum":1,"description":"Номер страницы","default":1,"title":"Page"},"description":"Номер страницы"},{"name":"size","in":"query","required":false,"schema":{"type":"integer","maximum":100,"minimum":1,"description":"Количество элементов на странице (1-100)","default":50,"title":"Size"},"description":"Количество элементов на странице (1-100)"}],"responses":{"200":{"description":"Объект с полями items и total","content":{"application/json":{"schema":{"$ref":"#/components/schemas/FilmListResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/films/search/":{"get":{"tags":["Кинопроизведения"],"summary":"Поиск кинопроизведений","description":"Полнотекстовый поиск по названиям и описаниям кинопроизведений","operationId":"search_films_api_v1_films_search__get","parameters":[{"name":"query","in":"query","required":true,"schema":{"type":"string","minLength":2,"description":"Поисковый запрос","title":"Query"},"description":"Поисковый запрос"},{"name":"page","in":"query","required":false,"schema":{"type":"integer","minimum":1,"description":"Номер страницы","default":1,"title":"Page"},"description":"Номер страницы"},{"name":"size","in":"query","required":false,"schema":{"type":"integer","maximum":100,"minimum":1,"description":"Количество элементов","default":50,"title":"Size"},"description":"Количество элементов"}],"responses":{"200":{"description":"Название и рейтинг фильма","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/FilmBaseResponse"},"title":"Response Search Films Api V1 Films Search  Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/genres/{genre_id}":{"get":{"tags":["Жанры"],"summary":"Получить информацию о жанре","description":"Возвращает полную информацию о жанре по его идентификатору","operationId":"genre_details_api_v1_genres__genre_id__get","parameters":[{"name":"genre_id","in":"path","required":true,"schema":{"type":"string","description":"UUID жанра","title":"Genre Id"},"description":"UUID жанра","example":"3d8d9bf5-0d90-4353-88ba-4ccc5d2c07ff"}],"responses":{"200":{"description":"Объект с информацией о жанре","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Genre"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/genres/":{"get":{"tags":["Жанры"],"summary":"Получить список жанров","description":"Возвращает список жанров с их уникальными идентификаторами","operationId":"list_genres_api_v1_genres__get","parameters":[{"name":"page","in":"query","required":false,"schema":{"type":"integer","minimum":1,"description":"Номер страницы","default":1,"title":"Page"},"description":"Номер страницы"},{"name":"size","in":"query","required":false,"schema":{"type":"integer","maximum":100,"minimum":1,"description":"Количество элементов на странице (1-100)","default":50,"title":"Size"},"description":"Количество элементов на странице (1-100)"}],"responses":{"200":{"description":"Список жанров","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Genre"},"title":"Response List Genres Api V1 Genres  Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/persons/{person_id}":{"get":{"tags":["Персоны"],"summary":"Получить информацию о персоне","description":"Возвращает информацию о персоне по его идентификатору","operationId":"person_details_api_v1_persons__person_id__get","parameters":[{"name":"person_id","in":"path","required":true,"schema":{"type":"string","description":"UUID персоны","title":"Person Id"},"description":"UUID персоны","example":"5b4bf1bc-3397-4e83-9b17-8b10c6544ed1"}],"responses":{"200":{"description":"Объект с информацией о персоне","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PersonFullResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/persons/{person_id}/films":{"get":{"tags":["Персоны"],"summary":"Получить список фильмов персоны","description":"Возвращает список из короткой информации по кинопроизведениям персоны","operationId":"list_films_api_v1_persons__person_id__films_get","parameters":[{"name":"person_id","in":"path","required":true,"schema":{"type":"string","description":"UUID персоны","title":"Person Id"},"description":"UUID персоны","example":"5b4bf1bc-3397-4e83-9b17-8b10c6544ed1"}],"responses":{"200":{"description":"Список кинопроизведений персоны с короткой информацией","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/PersonFilms"},"title":"Response List Films Api V1 Persons  Person Id  Films Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/persons/search/":{"get":{"tags":["Персоны"],"summary":"Поиск персон","description":"Полнотекстовый поиск персон по имени","operationId":"search_films_api_v1_persons_search__get","parameters":[{"name":"query","in":"query","required":true,"schema":{"type":"string","minLength":2,"description":"Поисковый запрос","title":"Query"},"description":"Поисковый запрос"},{"name":"page","in":"query","required":false,"schema":{"type":"integer","minimum":1,"description":"Номер страницы","default":1,"title":"Page"},"description":"Номер страницы"},{"name":"size","in":"query","required":false,"schema":{"type":"integer","maximum":100,"minimum":1,"description":"Количество элементов","default":50,"title":"Size"},"description":"Количество элементов"}],"responses":{"200":{"description":"Роли и фильмы персон","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/PersonFullResponse"},"title":"Response Search Films Api V1 Persons Search  Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"FilmBaseResponse":{"properties":{"uuid":{"type":"string","title":"Uuid","description":"Уникальный идентификатор фильма","example":"3d825f60-9fff-4dfe-b294-1a45fa1e115d"},"title":{"type":"string","maxLength":255,"title":"Title","description":"Название кинопроизведения","example":"Star Wars: Episode IV - A New Hope"}},"type":"object","required":["uuid","title"],"title":"FilmBaseResponse","description":"Базовая информация по кинопроизведениям"},"FilmFullResponse":{"properties":{"uuid":{"type":"string","title":"Uuid","description":"Уникальный идентификатор фильма","example":"3d825f60-9fff-4dfe-b294-1a45fa1e115d"},"title":{"type":"string","maxLength":255,"title":"Title","description":"Название кинопроизведения","example":"Star Wars: Episode IV - A New Hope"},"imdb_rating":{"anyOf":[{"type":"number","maximum":10.0,"minimum":0.0},{"type":"null"}],"title":"Imdb Rating","description":"Рейтинг кинопроизведения на IMDB","example":8.6},"genres":{"anyOf":[{"items":{"$ref":"#/components/schemas/Genre"},"type":"array"},{"type":"null"}],"title":"Genres","description":"Список жанров кинопроизведения"},"description":{"anyOf":[{"type":"string","maxLength":2000},{"type":"null"}],"title":"Description","description":"Описание кинопроизведения","example":"The Imperial Forces..."},"pg_rating":{"anyOf":[{"type":"string","maxLength":10},{"type":"null"}],"title":"Pg Rating","description":"Возрастной рейтинг (PG, PG-13, R и т.д.)","example":"PG-13"},"duration":{"anyOf":[{"type":"integer","exclusiveMinimum":0.0},{"type":"null"}],"title":"Duration","description":"Длительность фильма в минутах","example":120},"release_year":{"anyOf":[{"type":"integer","maximum":2031.0,"minimum":1888.0},{"type":"null"}],"title":"Release Year","description":"Год выпуска кинопроизведения","example":1977},"poster_url":{"anyOf":[{"type":"string","maxLength":2083,"minLength":1,"format":"uri"},{"type":"null"}],"title":"Poster Url","description":"URL постера фильма","example":"https://example.com/posters/star_wars.jpg"},"actors":{"anyOf":[{"items":{"$ref":"#/components/schemas/Person"},"type":"array"},{"type":"null"}],"title":"Actors","description":"Список актёров"},"writers":{"anyOf":[{"items":{"$ref":"#/components/schemas/Person"},"type":"array"},{"type":"null"}],"title":"Writers","description":"Список сценаристов"},"directors":{"anyOf":[{"items":{"$ref":"#/components/schemas/Person"},"type":"array"},{"type":"null"}],"title":"Directors","description":"Список режиссёров"}},"type":"object","required":["uuid","title"],"title":"FilmFullResponse","description":"Полная информация по кинопроизведениям"},"FilmListResponse":{"properties":{"items":{"items":{"$ref":"#/components/schemas/FilmShortResponse"},"type":"array","title":"Items","default":[]},"total":{"type":"integer","minimum":0.0,"title":"Total"}},"type":"object","required":["total"],"title":"FilmListResponse","description":"Список фильмов страницы и общее число записей по фильтру."},"FilmShortResponse":{"properties":{"uuid":{"type":"string","title":"Uuid","description":"Уникальный идентификатор фильма","example":"3d825f60-9fff-4dfe-b294-1a45fa1e115d"},"title":{"type":"string","maxLength":255,"title":"Title","description":"Название кинопроизведения","example":"Star Wars: Episode IV - A New Hope"},"imdb_rating":{"anyOf":[{"type":"number","maximum":10.0,"minimum":0.0},{"type":"null"}],"title":"Imdb Rating","description":"Рейтинг кинопроизведения на IMDB","example":8.6},"genres":{"anyOf":[{"items":{"$ref":"#/components/schemas/Genre"},"type":"array"},{"type":"null"}],"title":"Genres","description":"Список жанров кинопроизведения"}},"type":"object","required":["uuid","title"],"title":"FilmShortResponse","description":"Короткая информация по кинопроизведениям"},"Genre":{"properties":{"uuid":{"type":"string","title":"Uuid","description":"Уникальный идентификатор жанра","example":"3d8d9bf5-0d90-4353-88ba-4ccc5d2c07ff"},"name":{"type":"string","maxLength":50,"title":"Name","description":"Название жанра","example":"Action"},"description":{"anyOf":[{"type":"string","maxLength":1000},{"type":"null"}],"title":"Description","description":"Описание жанра","example":"Фильмы с динамично развивающимся сюжетом"}},"type":"object","required":["uuid","name"],"title":"Genre","description":"Модель жанра фильма"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"Person":{"properties":{"uuid":{"type":"string","title":"Uuid","description":"Уникальный идентификатор персоны","example":"5b4bf1bc-3397-4e83-9b17-8b10c6544ed1"},"full_name":{"type":"string","maxLength":255,"title":"Full Name","description":"Полное имя персоны","example":"George Lucas"}},"type":"object","required":["uuid","full_name"],"title":"Person","description":"Модель персоны (актёра, режиссёра, сценариста)"},"PersonFilms":{"properties":{"uuid":{"type":"string","title":"Uuid","description":"Уникальный идентификатор фильма","example":"3d825f60-9fff-4dfe-b294-1a45fa1e115d"},"title":{"type":"string","maxLength":255,"title":"Title","description":"Название кинопроизведения","example":"Star Wars: Episode IV - A New Hope"},"imdb_rating":{"anyOf":[{"type":"number","maximum":10.0,"minimum":0.0},{"type":"null"}],"title":"Imdb Rating","description":"Рейтинг кинопроизведения на IMDB","example":8.6}},"type":"object","required":["uuid","title"],"title":"PersonFilms","description":"Модель фильмов с рейтингом"},"PersonFullResponse":{"properties":{"uuid":{"type":"string","title":"Uuid","description":"Уникальный идентификатор персоны","example":"5b4bf1bc-3397-4e83-9b17-8b10c6544ed1"},"full_name":{"type":"string","maxLength":255,"title":"Full Name","description":"Полное имя персоны","example":"George Lucas"},"films":{"anyOf":[{"items":{},"type":"array"},{"type":"null"}],"title":"Films","description":"Список кинопроизведений с должностями человека"}},"type":"object","required":["uuid","full_name"],"title":"PersonFullResponse","description":"Полная информация по персоне"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}},"securitySchemes":{"HTTPBearer":{"type":"http","scheme":"bearer"}}}}